+++ ide/vscode/skip/src/extension.ts
@@ -1,0 +1,33 @@
+import * as vscode from "vscode";
+import * as which from "which";
+import { execSync } from "child_process";
+
+function formatSkipFile(fileName: string): void {
+  const skfmt = which.sync("skfmt");
+  execSync(`"${skfmt} -i < "${fileName}"`, { stdio: "inherit" });
+}
+
+export function activate(context: vscode.ExtensionContext) {
+  console.log("Skip extension is alive!");
+
+  const disposable = vscode.languages.registerDocumentFormattingEditProvider(
+    "skip",
+    {
+      provideDocumentFormattingEdits(
+        document: vscode.TextDocument,
+      ): vscode.TextEdit[] {
+        vscode.window.showInformationMessage("Hello World from skip!");
+
+        formatSkipFile(document.fileName);
+
+        return [];
+      },
+    },
+  );
+
+  context.subscriptions.push(disposable);
+}
+
+export function deactivate() {
+  console.log("Shutting down Skip extension");
+}