+++ ide/vscode/skip/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Change Log
 
-## Unreleased
+## 0.0.2
 
+- Formatting using `skfmt`.
+
+## 0.0.1
+
+- Syntax highlighting.